Implements run to line.
